+++ CHANGES.md
@@ -3,7 +3,9 @@
 ## 0.11.0
 
 - **BREAKING CHANGES: Renamed `is_band` to `is_sat_band` to better reflect that this function only checks optical and SAR bands**
+- **ENH: Adding a CustomProduct, allowing the user to load any stack as an EOReader Product !**
 - **ENH: Check if band exists before trying to load it**
+
 ## 0.10.2 (2022-MM-DD)
 
 - DOC: Update the installation paragraph in README